chore: disable dependabot
It was defunct because alerts were ignored or something. The last
dependabot PR was merged in 2023.

For a static website hosting exclusively public information, the impact
of any known CVEs is low to non-existent. The automation adds a lot
of churn and change risk in contrast to little added value.

It makes more sense to periodically manually updated packages with
docusaurus version upgrades, which occassionally contain breaking
changes that we need to manually review anyway.